Metric: Log extended smp pairing events
Include transport and complete information in extended enums.

Tag: #refactor
Bug: 236677941
Test: atest gd/cert/run

+++ system/test/mock/mock_main_shim_metrics_api.cc
@@ -128,7 +128,7 @@ void bluetooth::shim::LogMetricReadTxPowerLevelResult(
       raw_address, handle, cmd_status, transmit_power_level);
 }
 void bluetooth::shim::LogMetricSmpPairingEvent(
-    const RawAddress& raw_address, uint8_t smp_cmd,
+    const RawAddress& raw_address, uint16_t smp_cmd,
     android::bluetooth::DirectionEnum direction, uint8_t smp_fail_reason) {
   mock_function_count_map[__func__]++;
   test::mock::main_shim_metrics_api::LogMetricSmpPairingEvent(
